Rework the caching of images during install
According to the documentation, there's no need for clusterOSimage.
Which leaves us with the bootstrapOSimage, and all the installer does
is download it from the given url and decompress it to
.cache/openshift-installer/image_cache

We're gonna bypass running a container to serve that image over http
and just download and decompress in-place for the installer.