Fix openshift_install_version
lookups happen on localhost, and the tools on localhost do not reflect
the tools on the provisioner. So instead, we run the command on the
provisioner after we've installed the correct tools there. Then we
regex_search and set a fact for _openshift_install_version.

Note the prefixed `_`, it's a good convention that variables not intended
to be overridden are prefixed with underscore.