Adjust MTU for IPsec
https://github.com/openshift/cluster-network-operator/pull/1778

I'm not certain about this. The original documentation had 1352,
which was 2 bytes less than the overhead calculation. But if we
subtract for padding/alignment on 4/8 that would account for these
missing 2 bytes. (I'm picking 8 because we want to support ipv6,
but the end result is the same MTU size for 4 and 8 alignment)

From the rfc4303

2.3. Payload Data

   Note that the beginning of the next layer protocol header MUST be
   aligned relative to the beginning of the ESP header as follows.  For
   IPv4, this alignment is a multiple of 4 bytes.  For IPv6, the
   alignment is a multiple of 8 bytes.