okd-provisioner: Improve get-release-version
We always get the latest stable ocp version, for tooling.

If we're deploying OCP, the release_pull has to be a digest.
If we haven't specified a release_version, we use the latest stable.

If we're deploying OKD and haven't specified a version, we get the
latest stable.

The rest takes care of itself.