Split into two use-cases for mirror-registry
There's the installation of mirror-registry, and there's the usage
of a mirror-registry. And since we can use a "bigger" local registry,
we don't need to install mirror-registry anymore. And since it doesn't
operate on port 8443, we need to adapt somewhat.

Should probably make the mirrors/ org configurable at some point.